+++ server/src-test/Hasura/RQL/WebhookTransformsSpec.hs
@@ -1,11 +1,11 @@
-module Hasura.RQL.RequestTransformSpec (spec) where
+module Hasura.RQL.WebhookTransformsSpec (spec) where
 
 import Data.Aeson
 import Data.CaseInsensitive qualified as CI
 import Data.List (nubBy)
 import Data.Set qualified as S
 import Hasura.Prelude
-import Hasura.RQL.DDL.RequestTransform
+import Hasura.RQL.DDL.WebhookTransforms
 import Hedgehog.Gen qualified as Gen
 import Hedgehog.Range qualified as Range
 import Test.Hspec
@@ -29,16 +29,16 @@ spec = do
     hedgehog $ do
       headers <- forAll genTransformHeaders
       let sortH TransformHeaders {..} = TransformHeaders (sort addHeaders) (sort removeHeaders)
-          headersMaybe = decode $ encode headers
-      Just (sortH headers) === fmap sortH headersMaybe
+          headersMaybe = eitherDecode $ encode headers
+      Right (sortH headers) === fmap sortH headersMaybe
 
-  it "MetadataTransform RoundTrip" $
+  it "MetadataRequestTransform RoundTrip" $
     hedgehog $ do
-      transform <- forAll genMetadataTransform
+      transform <- forAll genMetadataRequestTransform
       let sortH TransformHeaders {..} = TransformHeaders (sort addHeaders) (sort removeHeaders)
-          sortMT mt@MetadataTransform {mtRequestHeaders, mtQueryParams} = mt {mtRequestHeaders = sortH <$> mtRequestHeaders, mtQueryParams = sort <$> mtQueryParams}
-          transformMaybe = decode $ encode transform
-      Just (sortMT transform) === fmap sortMT transformMaybe
+          sortMT mt@MetadataRequestTransform {mtRequestHeaders, mtQueryParams} = mt {mtRequestHeaders = sortH <$> mtRequestHeaders, mtQueryParams = sort <$> mtQueryParams}
+          transformMaybe = eitherDecode $ encode transform
+      Right (sortMT transform) === fmap sortMT transformMaybe
 
 trippingJSON :: (Show a, Eq a, ToJSON a, FromJSON a, MonadTest m) => a -> m ()
 trippingJSON val = tripping val (toJSON) (fromJSON)
@@ -56,6 +56,9 @@ genTemplateText = TemplateText . wrap <$> Gen.text (Range.constant 3 20) Gen.alp
   where
     wrap txt = "\"" <> txt <> "\""
 
+genStringTemplateText :: Gen StringTemplateText
+genStringTemplateText = StringTemplateText <$> Gen.text (Range.constant 3 20) Gen.alphaNum
+
 genContentType :: Gen ContentType
 genContentType = Gen.enumBounded @_ @ContentType
 
@@ -64,7 +67,7 @@ genTransformHeaders = do
   numHeaders <- Gen.integral $ Range.constant 1 20
 
   let genHeaderKey = CI.mk <$> Gen.text (Range.constant 1 20) Gen.alphaNum
-      genHeaderValue = coerce <$> genTemplateText
+      genHeaderValue = genStringTemplateText
 
       genKeys = S.toList <$> Gen.set (Range.singleton numHeaders) genHeaderKey
       genValues = S.toList <$> Gen.set (Range.singleton numHeaders) genHeaderValue
@@ -76,8 +79,8 @@ genTransformHeaders = do
 genQueryParams :: Gen [(StringTemplateText, Maybe StringTemplateText)]
 genQueryParams = do
   numParams <- Gen.integral $ Range.constant 1 20
-  let keyGen = coerce <$> genTemplateText
-      valueGen = Gen.maybe $ coerce <$> genTemplateText
+  let keyGen = genStringTemplateText
+      valueGen = Gen.maybe $ genStringTemplateText
   keys <- Gen.list (Range.singleton numParams) keyGen
   values <- Gen.list (Range.singleton numParams) valueGen
   pure $ nubBy (\a b -> fst a == fst b) $ zip keys values
@@ -88,8 +91,8 @@ genUrl = do
 
   pure $ StringTemplateText $ "http://www." <> host <> ".com"
 
-genMetadataTransform :: Gen MetadataTransform
-genMetadataTransform = do
+genMetadataRequestTransform :: Gen MetadataRequestTransform
+genMetadataRequestTransform = do
   method <- Gen.maybe genRequestMethod
   -- NOTE: At the moment no need to generate valid urls or templates
   -- but such instances maybe useful in the future.
@@ -98,7 +101,7 @@ genMetadataTransform = do
   contentType <- Gen.maybe $ genContentType
   queryParams <- Gen.maybe $ genQueryParams
   reqHeaders <- Gen.maybe $ genTransformHeaders
-  MetadataTransform
+  MetadataRequestTransform
     method
     url
     bodyTransform